+++ src/main/java/vendingmachine/Application.java
@@ -1,7 +1,6 @@
 package vendingmachine;
 
 import vendingmachine.controller.VendingMachineController;
-import vendingmachine.domain.MerchandiseList;
 
 public class Application {
     public static void main(String[] args) {
